Fix nullish support for older PSL versions and expand CI matrix
- Use string literal instead of NullishType::class to avoid class-not-found
  errors on PSL versions without NullishType
- Move nullish tests to separate fixtures, gated by class_exists()
- Add PHP 8.2, 8.3, 8.4, 8.5 to CI matrix
