+++ tests/Type/PslTypeSpecifyingExtensionTest.php
@@ -17,6 +17,11 @@ class PslTypeSpecifyingExtensionTest extends TypeInferenceTestCase
 		yield from $this->gatherAssertTypes(__DIR__ . '/data/coerce.php');
 		yield from $this->gatherAssertTypes(__DIR__ . '/data/assert.php');
 		yield from $this->gatherAssertTypes(__DIR__ . '/data/matches.php');
+		if (class_exists(\Psl\Type\Internal\NullishType::class)) {
+			yield from $this->gatherAssertTypes(__DIR__ . '/data/nullishCoerce.php');
+			yield from $this->gatherAssertTypes(__DIR__ . '/data/nullishAssert.php');
+			yield from $this->gatherAssertTypes(__DIR__ . '/data/nullishMatches.php');
+		}
 		if (InstalledVersions::satisfies(new VersionParser(), 'azjezz/psl', '<2.0.0')) {
 			yield from $this->gatherAssertTypes(__DIR__ . '/data/complexTypev1.php');
 		} else {
