+++ tests/Type/data/coerce.php
@@ -29,35 +29,6 @@ class GeneralTest
         assertType('array<mixed>', $input);
     }
 
-    /**
-     * @param array<mixed> $input
-     */
-    public function coerceNullishShape(array $input): void
-    {
-        $specification = Type\shape([
-            'name' => Type\string(),
-            'bio' => Type\nullish(Type\string()),
-        ]);
-
-        $output = $specification->coerce($input);
-
-        assertType('array{name: string, bio: string|null}', $output);
-    }
-
-    /**
-     * @param array<mixed> $input
-     */
-    public function coerceOptionalNullishShape(array $input): void
-    {
-        $specification = Type\shape([
-            'bio' => Type\optional(Type\nullish(Type\string())),
-        ]);
-
-        $output = $specification->coerce($input);
-
-        assertType('array{bio?: string|null}', $output);
-    }
-
 	public function coerceInt($i): void
 	{
 		$spec = Type\int();
